user and admin can add games
