+++ backend/src/main/java/com/example/dontwastetomuch/game/GameRepo.java
@@ -3,7 +3,7 @@ package com.example.dontwastetomuch.game;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+
 
 @Repository
 public interface GameRepo extends MongoRepository<Game, String> {
