+++ frontend/src/Pages/RegisterPage.tsx
@@ -1,0 +1,10 @@
+
+export default function RegisterPage(){
+    return(
+        <div>
+            <input type={"text"}/>
+            <input type={"password"}/>
+            <button>Login</button>
+        </div>
+    )
+}
