+++ frontend/src/Pages/WelcomePage.tsx
@@ -1,0 +1,10 @@
+
+export default function WelcomePage(){
+    return(
+        <div>
+            <input type={"text"}/>
+            <input type={"password"}/>
+            <button>Login</button>
+        </div>
+    )
+}
