+++ frontend/src/components/GameCard.tsx
@@ -1,0 +1,7 @@
+
+
+export default function GameCard(){
+    return(
+        null
+    )
+}
